requirements.txt
qdrant-client>=1.7.0
openai>=1.0.0
sentence-transformers>=2.2.0
torch
numpy
tqdm
python-dotenv
colorama
jieba
rank_bm25
networkx
pypinyin
fastapi
uvicorn
websockets
requests
aiohttp
aiosqlite
loguru>=0.7.0
